\name{sample_initial}
\alias{sample_initial}
\alias{sample_subtracted}
\alias{sample_cluster}
\alias{sample_goterms}
\alias{sample_complete}
\alias{sample_cluster}
\alias{control_initial}
\docType{data}
\title{
Sample Objects for the msarc Package
}
\description{
These \code{msarc} objects are used as examples in the \code{msarc} package.
}
\usage{data(sample_initial)}
\format{
  These are \code{msarc} objects.
}
\source{
From Mohammed, H., D'Santos, C., Serandour, A. A., Ali, H. R., Brown, G. D., Atkins, A., et al. (2013). Endogenous Purification Reveals GREB1 as a Key Estrogen Receptor Regulatory Factor. Cell reports, 3(2), 342-349. doi:10.1016/j.celrep.2013.01.010
}
\examples{
data(sample_initial,package="msarc")
data(control_initial,package="msarc")
}
\keyword{datasets}
